Add pagination to transactions

+++ monlixv2/src/main/java/com/monlixv2/service/models/transactions/TransactionResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class TransactionResponse(
     @SerializedName("ptcClicks")
-    val ptcClicks: Double,
+    val ptcClicks: Int,
     @SerializedName("ptcEarnings")
     val ptcEarnings: Double,
     @SerializedName("completedOffers")
